Add Android build to POM
1. jCardSim currently uses the Maven *shade* plugin to include
classes from *bouncycastle* into the jCardSim JAR.  This causes
class loader problems for Android users and may also cause problems
for JDK users that want to use a different *bouncycastle* version.
Therefore we change the *shade* plugin configuration to re-package
*org.bouncycastle* to *com.licel.jcardsim.bouncycastle*.

2. Android does not provide the packages *java.rmi* and
*javax.smartcardio*.  Therefore we add a *shade* configuration that
creates an additional JAR file: *jcardsim-VERSION-android.jar*.
This JAR file does not contain classes that depend on *java.rmi*
or *javax.smartcardio*.  To ensure that the JAR file still works we
use a script that runs all applicable unit tests against the JAR.
Since this is very hard to do in "plain maven" we use *antrun* to
implement the script.