Update instructions for building and development
Re-word build instructions for consistency and clarity. Mention that
jCardSim can only be built under Windows. Include the installation
of JDK 8 as the first step. The user's Maven settings no longer need
to be modified to locate the JCDK. Mention the JAVA_HOME environment
variable, which must be set for Maven to work but is not done by the
Windows installer.

Move the opening comment about contributing to jCardSim underneath
the "Development" section. The main goal of this page is to provide
instructions for building jCardSim, rather than for contributing to
its development.