Improve documentation for the API
Refine the wording used to describe the API and the different ways
it can be called. There is only one (Java-based) "API" to jCardSim,
just as there is only one Java Card API. The different ways to use
this API can be more clearly distinguished by the name of the class
or interface that is the means for interacting with the simulator:
CardSimulator, CardTerminalSimulator, or JavaCardRemoteInterface.

For clarity, refer to classes in javax.smartcardio using the fully-
qualified class name. This distinguishes them from classes provided
by jCardSim.

Copy the brief description of HelloWorldApplet's behavior from the
quick start guide for the CLI (and slightly tweak it).

When mentioning the CLI configuration file, link to the sample file
in the respository.