Improve documentation for the CLI
Focus only on the CLI. Remove statements about "choosing" between
the API and the CLI, or which consider this as the first part of a
two-part series; they are not helpful. Instead, begin with a high-
level summary of the classes that can be used to run the simulator
from the command line.

Include BixVReaderCard as part of the CLI. For now, simply link to
the blog post which describes how to use it.

Move a statement about jCardSim working directly with class files
down the page to where this detail is most relevant.

Remove a direct link to download a particular version of jCardSim.
It should be downloaded from a different page on the website which
may link to more than one version.